Data Dictionary/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
aiomysql==0.2.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
loguru==0.7.2
pgvector==0.2.4
psycopg[binary]==3.1.12
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
# slowapi not required any more due to middleware fallbacks

